fix(bracelet): changed Serendipity type to decimal

+++ CLOFT.SerenUp/CLOFT.SerenUp.Domain/Models/Bracelet.cs
@@ -7,7 +7,7 @@ public class Bracelet
 {
     public Guid SerialNumber { get; set; }
     public string? Username { get; set; }
-    public int? Serendipity { get; set; }
+    public decimal? Serendipity { get; set; }
     public string Color { get; set; }
 
     public Bracelet()
